Moved the function for processFile into a separate file from the script that is actually executed when a worker thread is forked so that it can be tested reliably
